FIX: Se ha incluido nuevas funciones para no saturar el script.js, se empezara a modularizar para ello
