+++ cliente/helper/index.js
@@ -1,2 +1,5 @@
 export * from './get.js'
 export * from './post.js'
+export * from './patch.js'
+export * from './delete.js'
+export * from './utils.js'
